refactor(solver): rehome tail-call instantiation into recursive_growth module
Move try_instantiate_application_for_tail_call out of the oversized evaluate.rs
into the under-cap recursive_growth module alongside the divergence detector,
leaving only narrow call-site glue at evaluate_application. Net effect: evaluate.rs
shrinks below its main size while the TS2589 divergence-guard behavior is
unchanged.
